Made a few pretty major improvements:
(1) changed SQP convergence check so that it is checking the dual
residual only for the *nonzero entries* of the current solution,
x. Also, add a more informative comment to accompany this check.

(2) Change the b.min() >= -convtolactiveset stopping criterion so that
it is only over co-ordinates in the working set (that is, the "zero"
co-ordinates).

(3) Fix newind = b.index_min() so that it is limited to the active
(working) set.